+++ server/app/sqls/connection.py
@@ -1,0 +1,14 @@
+import os
+
+import psycopg2
+from dotenv import load_dotenv
+
+load_dotenv()
+
+
+def get_connection():
+    host = os.environ.get("PGHOST")
+    dbname = os.environ.get("PGDATABASE")
+    user = os.environ.get("PGUSER")
+    password = os.environ.get("PGPASSWORD")
+    return psycopg2.connect(host=host, dbname=dbname, user=user, password=password)
